Define in-game screen size to 480x270 voxels
This means that, at most, the screen will render 480x270 voxels of the
game map. If you use a larger resolution, the size of a pixel will
increase so that you still see the same portion of the game map. This
number was defined after testing, as it seems to be the largest size on
which the game performance stays reasonable in low-end computers.