Lazy model loader, remove docs, minify
The compiled binary was growing huge (>40MB), causing the site to take a
logn time to load. Now, the lazy model loader only downloads a model
when it is used in game.

The docs folder since we don't use github.io anymore, and it was
bloating the repository history.

Minification was enabled on webpack.config.js.


Former-commit-id: 4c205131f172e05c46ddd63f7dcee3ca387670ac