New CPU mode, site layout, game stage and camera
The CPU mode doesn't use ray marching anymore, but, instead, just draws
the voxels directly to the screen with a depth buffer. That makes it
much faster, but it won't be able to have advanced light effects and
will, thus, produce different results.

The size of the game stage was reduced from 1024x1024 to 896x384. It is
stretched to fit the screen, so a camera isn't needed anymore. The site
layout was modified accourdingly. The game occupies the center of the
site. The chat now is on left/bottom of the UI, and a new control panel
was added next to it.